♻️ Changed applicable `timestamp`s from `u64` to `u128` (#961)
fixes: #951

- **chore: timestamp types `u64` > `u128`**

I changed all the applicable `timestamp`s and related functions from
`u64` to `u128`, with an exception being any originating from `as_secs`
or other std sources as they're `u64` by default.

+++ crates/hyle-model/src/utils.rs
@@ -10,15 +10,15 @@ pub fn get_current_timestamp() -> u64 {
         .as_secs()
 }
 
-pub fn get_current_timestamp_ms() -> u64 {
+pub fn get_current_timestamp_ms() -> u128 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .expect("Time went backwards")
-        .as_millis() as u64
+        .as_millis()
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, BorshDeserialize, BorshSerialize)]
-pub struct TimestampMs(pub u64);
+pub struct TimestampMs(pub u128);
 
 impl TimestampMs {
     pub fn now() -> TimestampMs {
